Post a new food to the table
