+++ foods.html
@@ -17,11 +17,10 @@
     <input type="text" name="new-food-name" id="new-food-name"><br>
     Calorie Count:<br>
     <input type="text" name="new-food-calories" id="new-food-calorie-count"><br>
-    <input type="submit" value="Add Food" id="new-food-button">
+    <button id="new-food-button">Add Foood</button>
   </form>
 
   <script src="http://ajax.googleapis.com/ajax/libs/jquery/1.7.1/jquery.min.js" type="text/javascript"></script>
-
   <script src="main.bundle.js"></script>
 </body>
 </html>
